app/backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
flask_apscheduler==1.13.1
python-dotenv==1.0.1
gunicorn==21.2.0
mysql-connector-python==8.0.33
flask-jwt-extended==4.5.3
pyjwt==2.8.0
bcrypt==4.0.1
boto3==1.38.13
python-magic==0.4.27